requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.3
django-admin-tailwind==0.0.16
django-ckeditor==6.4.1
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-taggit==2.1.0
django-tinymce==3.4.0
django-utils-six==2.0
gunicorn==20.1.0
idna==3.3
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-decouple==3.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0